# perf(ext/cache): set journal_mode=wal